pyproject.toml
[tool.poetry]
name = "freelance-django"
version = "0.1.0"
description = ""
authors = ["Safaraliyev05 <safaraliyevsardor05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
django-jazzmin = "^3.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
